fest(sql): Refactor handler
